feat(DTFS2-7556): add createTfmUser test helper
